backend/requirements/base.txt
Django==1.8.4
django-extensions==1.5.5
django-model-utils==2.3.1
django-ordered-model==0.4.2
django-filebrowser==3.5.7
django-grappelli==2.7.1
django-ordered-model==0.4.2
djangorestframework==3.2.2
django-cors-headers==1.1.0
Pillow==2.9.0

ipython
pytz==2015.4
